Simple cleaning of OIDC4VCI without changing the flow
