+++ lib/polygon_id/cubit/polygon_id_cubit.dart
@@ -559,7 +559,7 @@ class PolygonIdCubit extends Cubit<PolygonIdState> {
     await credentialsCubit.insertCredential(
         credential: credentialModel,
         blockchainType: walletCubit.state.currentAccount!.blockchainType,
-        uri: uri);
+        uri: uri,);
   }
 
   /// getSchemas
