+++ packages/oidc4vc/lib/src/functions/list_to_string.dart
